Update documentation for host_authorization plugin
The host_authorization plugin by design does not check the Host
header (or X-Forwarded-Host header) contains a host in a valid
format.  There is in general no need as exact checks are done
against the full hostname in the string case (the most common
case).

For regexps, the regexp provided by the user is used verbatim,
so we can only provide recommendations to the user for what to
use.  Similarly for the proc case.

While here, provide more detail about the :check_forwarded behavior.