Fix issue where loading content_security_policy plugin after default_headers plugin had no effect
An invalid optimization in the default_headers plugin prevented
the content_security_policy plugin from taking effect if the
default_headers plugin was loaded first.

Only apply the default_headers optimization when freezing the
application, since after that point no changes can be made and
we can be sure the optimization is safe.

This can prevent the default_headers optimization from applying
in cases where it could theoretically be applied, such as when
the content_security_policy plugin is loaded before the
default_headers plugin.  However, I think that is acceptable.

This issue was reported by Laurent Guinotte.