Use RFC 5987 UTF-8 and ISO-8859-1 encoded filenames when using send_file and attachment in the sinatra_helpers plugin
In addition to RFC 5987 encoded filenames, check the default filename
for parameter for RFC 5987 valid characters, and replace invalid RFC
5987 chracter sequences with a single dash.  I think this is a better
approach than the previous, which just used Ruby's inspect and would
result in backslash characters which are at least not allowed in
Windows filenames.

This causes backwards compatibility issues if you were expecting
Roda to pass non-RFC 5987 attr-char characters attachment filenames
without modification.  I don't expect it to be an issue, but an
option to use the previous approach could be added if users of the
sinatra_helpers plugin desire it.

Sinatra itself doesn't support RFC 5987 encoded filenames yet, so
technically this reduces compatibility with Sinatra.