Automatically optimize r.on/r.is/r.get/r.post methods with a single string, String, Integer, or regexp argument
Calling these methods with a single argument is very common, and
this optimization can have a significant effect, up to about 50%
improvement in routing in the best case.

This adds an internal _optimized_matching plugin, and automatically
uses the plugin if freezing the class if the class isn't subclassed
and hasn't overridden any of the related methods.  This plugin
inlines some matching code for individual string, String, Integer,
and regexp matchers, to avoid the need to call the slower if_match
method to handle matching.  If any of the methods is called with
no arguments or more than single argument, or called with an argument
that the plugin doesn't handle, the plugin falls back to calling
the default code.