Extract named_routes plugin from multi_route plugin
This keeps the r.multi_route support in the multi_route plugin,
but moves the support for r.route to a named_routes plugin. It
has the multi_route plugin depend on the named_routes plugin.
This way you don't need to load the multi_route plugin if you
are just using the r.route support.

One reason for this split is that the hash_routes plugin does not
have a direct equivalent of just dispatching to a block for a named
route.  So if you wanted to replace the multi_route plugin with the
hash_routes plugin, it was more challenging if you were using
r.route. Now you can switch from multi_route to hash_routes and
named_routes.