Use faster OpenSSL::Digest instead of Digest for assets plugin SRI support
In Ruby 3, digest is generally slower than openssl/digest on common
platforms, because openssl/digest will use assembly optimized
routines.  This isn't the case in older versions of Ruby, because
digest would fallback to the openssl implementation if it was
available.  However, that feature was removed in Ruby 3, when
digest moved to being a default gem.