Support testing with LINT environment variable
With LINT set, Rack::Lint is used for all created apps, checking
that the applications follow the Rack SPEC. This found multiple
SPEC violations, which were recently fixed.  Not all specs can
be run with LINT, since they test behavior when Roda is used
with requests that violate SPEC, so some specs are skipped in
this mode.