Add :forward_response_headers option to middleware plugin
Setting the :forward_response_headers configuration option to true
causes the middleware to apply response headers modified in the route
block to the response of the forwarded call when the middleware didn't
route the request.

  class App < Roda
    plugin :middleware, forward_response_headers: true

    route do |r|
      r.get 'a' do
        # ...
      end

      # this will be merged into main app's response headers
      response.headers['Foo'] = 'Bar'
    end
  end

This is useful in context of Rodauth, which provides methods that are
expected to be called inside the route block outside of routing, some of
which can modify response headers. On example is setting `Set-Cookie` to
extend the remember period when logging in an account from remember
token.