Update json_parser.rb
sorry, I am not a native speaker, so the wording might not be perfect. but having something similar as a clue at hand, would had helped a lot. I had to google to find out how the plugin was supposed to be used. Though there is a hint in what the `:wrap` option does.