Make HTTP version handling compatible with Rack 3 SPEC
Add RodaRequest#http_version for determining the HTTP version in use.
Use this in the chunked, common_logger, and status_303 plugins.

For the status_303 plugin, allow use of the 303 status for HTTP/2
and higher versions.  Only call super (302) for HTTP/1.0 and HTTP/0.9
versions.

For Rack 3 compatibility, prefer SERVER_PROTOCOL since it is required
to be set in Rack 3.  Otherwise, prefer HTTP_VERSION, since that is
what is used previously.  Puma, the most common Ruby webserver, sets
HTTP_VERSION accurately, but SERVER_PROTOCOL is fixed at 1.1, so
switching to always using SERVER_PROTOCOL could result in incorrect
behavior when using Puma (I'm sure Puma will fix this in their changes
to support Rack 3).