Add render_coverage plugin, using tilt 2.1 features to allow for compiled templates in Ruby <3.2
Tilt uses eval by default to execute template code, which Ruby
has historically not supported coverage for.

Ruby 3.2 added the ability for coverage to work with evaled code,
but it may be a while before everyone is running that, and it
seems nice to allow for template coverage testing in earlier
Ruby versions. This uses a new feature I added to Tilt 2.1 to
store template code in files and load the code instead of
evaling it, so that template code supports coverage in older
Ruby versions.

Another advantage of using load instead of eval is it allows for
displaying verbose syntax warnings for compiled template code.

This changes the Render::TemplateMtimeWrapper API in non-compatible
ways.  Any external usage of this class needs to be updated (this
class was already documented as internal with an API that could
change at any time).