Add link_to plugin with link_to method for creating HTML links
This is modeled somewhat after Rails link_to, though it is significantly
simplified.  The third argument is an HTML attributes hash, which directly
uses the hash keys and values as HTML attribute keys and values, there is
no support for non-attribute options such as :data, :method, :remote,
:confirm, or :disable_with.

The second argument (href), can be a string for an existing path/URL,
a symbol for a named path in the path plugin, or instances of a class
registered with the path plugin.

HTML escaping is done on the path and attribute values, but not the body.
If you want the body to be escaped, it's probably easiest to escape it
using the h plugin.

My personal preference is to construct HTML tags like this manually in
the template, instead of using a helper method, but I have received
multiple requests for a link_to method in the past, often related to
easing migration from Rails.