Add Integer_matcher_max plugin for setting maximum integer value matched by the Integer matcher
Having a maximum limit can ensure that the yielded integer value
is in an expected range.  If you are passing the integer value
to any code that expects a value that can fit into a specific
integer size, not matching on input that is too large can help
avoid processing a request that will likely result in an
exception.

The default maximum value for the plugin is 2**63-1, the maximum
value of a signed 64-bit integer. Custom values can be configured
by passing an argument to the plugin.