Add typecast_params_sized_integers plugin for converting parameters to sized integers
This allows for converting parameters to integers only if the
integer is the range for a specific integer sized type.  This
is mostly useful when you will be passing the integer value to
lower level code that expects specific sizes, to avoid issues
or rescue errors earlier if the value provided is outside of range
for the underlying integer type.

By default, the plugin supports 8-bit, 16-bit, 32-bit, and 64-bit
conversions, but the supported versions are configurable with the
:sizes option.  You can also specify a :default_size option, in
which case the default methods used will be limited to a given size.