Add RodaRequest#unsupported_block_result
This method is called when the block result is not nil, false, or
a String.  This allows plugins that want to support additional
block types to not longer have to override block_result_body,
so that they do not have a performance impact on block results
that are nil, false, or String.