Remove Gemfile
This was added for use by Travis, but it is a bad idea to use
Gemfile in libraries due to the creation of a lock file.  You
always want to make sure you are testing the library against
the latest version of the installed dependencies to catch
regressions, and having a lockfile prevents automatically picking
up newer versions of installed dependencies.