Add permissions_policy plugin for setting Permissions-Policy header
This is similar to the content_security_policy plugin, with the
main difference being the formatting of the header value, and
how :all/:none are used to specify whether domains are allowed.

This uses the Permissions-Policy header and not the older
Feature-Policy header, which may seem strange if you look at
caniuse.com, where Feature-Policy is supported by 97% of
browsers, compared to Permissions-Policy with only 71% support.
However, this is misleading because neither Firefox nor Safari
support the Feature-Policy header, only the allow attribute on
iframes, and this plugin is for the header only.  The only
real compatibility loss is for a few Android browsers,
amounting to about 4% difference in support.  Considering that
everything will still work on those browsers (the lack of
header support fails open), I think it's better to directly
implement the Permissions-Policy header instead of implementing
Feature-Policy now and implicitly switching to
Permissions-Policy later.