Support :namespace option in hmac_paths plugin, allowing for easy per-user/per-group HMAC paths
For applications where the same path shows different content
depending on the logged in user, user A who was provided an
hmac path could give that path to user B who had not been
provided an hmac path, and that path would work for user B.
By using the :namespace option, you the HMAC paths are
namespace-specific, so a path for a namespace used for user
A would not be valid in a namespace for user B (assuming the
namespaces are different).

The majority of cases where HMAC path namespaces are useful
will based the namespace on some session value, so a
:namespace_session_key plugin option has been added, which
sets a default namespace used by both the hmac_path and
r.hmac_path methods, so the namespace doesn't need to be
specified manually per-call.