Remove development dependency on sassc
Nothing in Roda requires sass/sassc.  It was used for testing the
assets plugin and a few other plugins, but there is no need to use
it, the testing can be done with other tilt template engines.

sass went end of life over 5 years ago. sassc's last release and
commit were over 4 years ago.  The recommended replacement library is
sass-embedded, but that depends on google-protobuf, and I'd like
to remove dependencies and not add them.