Add :zstd option to public plugin to supplement it to serve zstd-compressed files with .zst extension
Chrome has supported zstd accept encoding in March.  It offers faster compression
and decompression than brotli, though brotli can win slightly on compression ratio.
If the :zstd option is used, zstd it is preferred to both brotli and gzip, and it uses
a .zst file extension.

This also adds an :encodings option to the plugin, for configuration of the order
in which the encodings are attempted, and allowing for the ability to support
encodings other than zstd, brotli, and gzip, or non-default extensions.

While here, do some refactoring and optimization:

* Precompute the regexps used for checking for accepted encodings, instead of
  generating a new regexp at runtime for each request for each encoding.

* Use match? instead of match if supported, since the captures are not needed.

* Only access env['HTTP_ACCEPT_ENCODING'] once, and don't try serving compressed
  files if it isn't present.

This changes the private public_serve_compressed method in a non-backwards
compatible way, moving the encoding check from it to public_serve_with.
Users who call the method directly should probably update their code.

This removes the setting of the :public_gzip and :public_brotli application
options, in favor of :public_encodings.  Any code directly accessing the
:public_gzip and :public_brotli application options will need to be updated.

This also supports the :zstd and :encodings options in the multi_public
plugin, since it uses the same options as the public plugin.