Make capture_erb plugin not break if String#capture is defined
It turns out Rails 4 defines Kernel#capture, so using the capture_erb
plugin in Roda 3.82+ and Rails 4 in the same process did not work
correctly.  Not sure anyone is doing that, but add a check that the
buffer is not a String instance.  Calling capture directly on the
buffer is designed for use with erubi/capture_block.