Evaluate class_matcher and symbol_matcher blocks in route-block context
While this has a change to break backwards compatibility, the new
behavior is far more useful.  I would guess most class_matcher and
symbol_matcher blocks operate solely on their arguments and other
local variables and constants in scope, and do not depend on being
evaluated in a specific context (which would likely be application
class level context for most applications).

This allows for code such as:

  class_matcher Post, Integer do |id|
    Post.where(:user_id=>session['user_id']).with_pk(id)
  end