Support :assume_fixed_locals render plugin option for better caching when all templates use fixed locals
Previously, rendering a template with locals would always allocate
an array for the compiled method cache key.  Before template fixed
locals support was added last release, this was necessary because
the compiled method depended on the locals provided.  If you are
sure you are always using fixed locals, you can use the template
directly as the cache key, since the same compiled method will
work.  In addition to providing faster cache lookup, this also
avoids duplicate compiled methods if rendering a compiled method
both with and without locals (which can happen when all locals
are optional, or when the template does not accept locals, but
is called with an empty locals hash instead of no locals).

This has view_options freeze the returned key in
_cached_template_method_key, instead of returning an unfrozen
key and expecting the render plugin to freeze it.