Escape newlines in common logger lines
This uses the same approach I used in Rack.  I don't consider this
a security issue, because expoiting it requires a broken server
that is including newlines in things that it shouldn't.  Attempts
to exploit it even in that case leave a clear trail in the log,
since there will be a partial line directly before an attempt to
inject a line.