Support returns: :buffer method and plugin option in capture_erb plugin
This is how the method already behaved if the erubi/capture_block
template engine was used.  This use of returns: :buffer fixes cases
where a template block ends in a conditional, and the conditional
returns nil.  This should probably have been the default behavior
from the beginning, but as some users may be depending on the
current default behavior, this doesn't break backwards compatibility.
However, I plan to change the the default behavior in Roda 4 (users
can opt-in to the old behavior if they want at that point).