Add each_part plugin, with a simpler method for using render_each with locals
This offers the same simplificaiton advantage over the render_each plugin
as the part plugin offers over the the render plugin.  The performance
advantage is less, as this cannot use ... for passing arguments, as it
needs to modify the hash passed to the compiled template method to add
the entry for the current element.