Remove use of rdoc/task in Rakefile
This is an unnecessary require in the vast majority of cases, slowing
down the Rakefile.

Switch to using a normal task.  In addition to being faster, I think
this is easier to understand.  This just generates the documentation
into www/public/rdoc.  Doesn't seem necessary to have a separate
task to generate into rdoc.