Support invalid_value_message for custom invalid value messages per-type in typecast_params plugin
Previously, the typecast_params plugin used the same missing parameter
error message to handle both the case where the parameter was not
provided, and for the case where the provided parameter was converted
to nil.  This separates the error handling, so you only get the
missing parameter error message if the parameter is actually missing,
and you get the invalid value message if it is present but is
converted to nil.

For the built-in types that can result in a provided parameter being
converted to nil, provide appropriate default invalid value messages
via the :invalid_value_message option to handle_type.  Allow
overriding invalid value messages for the builtin types via the
invalid_value_message method.

This changes the 4th argument of the process and process_arg private
methods from max_input_bytesize to type, which should be the type
symbol passed to handle_type.  The parameters are now required.
Direct callers of these private methods will need to adjust to the
new API.

Code rescuing Roda::RodaPlugins::TypecastParams::Error and handling
the reason may need to adjust to handling :invalid_value in cases
where it was handling :missing.

Add invalid value messages to the types registered by the
typecast_params_sized_integers plugin.