Optimize render_each and each_part template selection when freezing app with :assume_fixed_locals render plugin option
This simplifies the _cached_render_each_template_method private
method to avoid multiple conditionals. It would be nice to remove
the case statement as well, since if you are using fixed locals
with keyword arguments, it's unlikely you could provide an
argument value that isn't a string or symbol.  However, it's
possible, either with an object with a custom to_s method, or
when using a template accepting a keyword splat.