Optimize render_each and each_part default local selection on Ruby 3+/!Windows
This uses an allocationless and regexp-free approach for symbols
and strings that do not have slashes or periods. This is about 3x
faster.

This support requires Ruby 3 for the Symbol#name usage and does
not work correctly on Windows (where \ can be a file separator),
so fallback to the previous code in other cases.