Add view_subdir_leading_slash plugin for using view subdirectory unless template name starts with slash
The default view_options plugin behavior is to use a view
subdirectory unless the template name contains a slash. That can
works well in most cases.  However, when there are subdirectories
of that view subdirectory that also contain templates, it requires
repeating the view subdirectory in the template name. To handle
the cases where the view subdirectory contains subdirectories with
templates, the view_subdir_leading_slash plugin can be used to
keep template names for the nested view subdirectory DRYer, at the
expense of having to use a leading slash for template names outside
the view subdirectory.