Raise Roda::RodaPlugins::Sessions::CookieTooLarge for cookies where total cookie size is over 4K (not just cookie value)
Browsers consider the entire cookie, including name and attributes
in the 4K limit, not just the cookie value. Previously, Roda only
raised an error for the cookie value being over. This would miss
cases where the cookie value was close to the limit, and the name
and attributes pushed the value over the limit.

To handle this case, calculate the entire cookie size when setting
the cookie, and raise if the size of the entire cookie is too large.