Add map_matcher plugin, for matching next segment in request path to a hash key, yielding hash value
This is designed to be a more efficient approach of a typical pattern
for route metaprogramming. When dealing with many similar routes, it
is common to use a hash keyed by route segment, and then match against
an array of the keys:

```ruby
map = {
  'album' => Album,
  'artist' => Artist,
  ...
}.freeze

keys = map.keys.freeze

route do |r|
  r.on "type", keys do |key|
    value = map[key]
    # ...
  end
end
```

For large maps, this is less efficient, since the array matcher
checks whether each element is the next route segment. The
map_matcher plugin allows for:

```ruby
route do |r|
  r.on "type", map: map do |value|
    # ...
  end
end
```

This gets the next route segment and checks whether it is a hash
key, instead of iterating over the hash elements, so it is more
efficient. It also results in simpler code.